rename file to have correct name
