+++ extension/src/ResourceLocator.ts
